bot/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
prettytable==3.10.0
pyTelegramBotAPI==4.18.1
requests==2.32.3
schedule==1.2.2
urllib3==2.2.1
wcwidth==0.2.13
python-dotenv==1.0.1